fix: cleanup and renaming
